Embeddings Refactoring
Refactored embeddings into a separate class.

#132
