+++ Generellem.Tests/AzureOpenAIRagTests.cs
@@ -2,6 +2,7 @@
 using Azure.AI.OpenAI;
 
 using Generellem.Document.DocumentTypes;
+using Generellem.Embedding;
 using Generellem.Llm;
 using Generellem.Rag.AzureOpenAI;
 using Generellem.Repository;
@@ -16,9 +17,9 @@ namespace Generellem.Rag.Tests;
 public class AzureOpenAIRagTests
 {
     readonly Mock<IAzureSearchService> azSearchSvcMock = new();
-    readonly Mock<IDynamicConfiguration> configMock = new();
     readonly Mock<IDocumentHashRepository> docHashRepMock = new();
     readonly Mock<IDocumentType> docTypeMock = new();
+    readonly Mock<IEmbedding> embedMock = new();
     readonly Mock<ILogger<AzureOpenAIRag>> logMock = new();
     readonly Mock<LlmClientFactory> llmClientFactMock = new();
     readonly Mock<OpenAIClient> openAIClientMock = new();
@@ -33,15 +34,9 @@ public class AzureOpenAIRagTests
             .Setup(doc => doc.GetTextAsync(It.IsAny<Stream>(), It.IsAny<string>()))
             .ReturnsAsync("text content");
 
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIEndpointName])
-            .Returns("https://generellem");
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIEmbeddingName])
-            .Returns("generellem-embedding");
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIApiKey])
-            .Returns("generellem-key");
+        embedMock
+            .Setup(e => e.GetEmbeddingOptions(It.IsAny<string>()))
+            .Returns(new EmbeddingsOptions());
 
         embedding = new ReadOnlyMemory<float>(TestEmbeddings.CreateEmbeddingArray());
         List<EmbeddingItem> embeddingItems =
@@ -83,120 +78,7 @@ public class AzureOpenAIRagTests
             .ReturnsAsync(chunks);
 
         azureOpenAIRag = new AzureOpenAIRag(
-            azSearchSvcMock.Object, configMock.Object, docHashRepMock.Object, llmClientFactMock.Object, logMock.Object);
-    }
-
-    [Fact]
-    public async Task EmbedAsync_CallsGetEmbeddingsAsync()
-    {
-        TextProcessor.ChunkSize = 5000;
-        TextProcessor.Overlap = 100;
-
-        await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None);
-
-        openAIClientMock.Verify(
-            client => client.GetEmbeddingsAsync(It.IsAny<EmbeddingsOptions>(), It.IsAny<CancellationToken>()),
-            Times.AtLeastOnce());
-    }
-
-    [Fact]
-    public async Task EmbedAsync_WithNoOverlap_BreaksTextInto2Chunks()
-    {
-        int oldChunSize = TextProcessor.ChunkSize;
-        int oldOverlap = TextProcessor.Overlap;
-        TextProcessor.ChunkSize = 9;
-        TextProcessor.Overlap = 0;
-
-        await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None);
-
-        openAIClientMock.Verify(
-            client => client.GetEmbeddingsAsync(It.IsAny<EmbeddingsOptions>(), It.IsAny<CancellationToken>()),
-            Times.Exactly(2));
-        TextProcessor.ChunkSize = oldChunSize;
-        TextProcessor.Overlap = oldOverlap;
-    }
-
-    [Fact]
-    public async Task EmbedAsync_WithOverlap_BreaksTextInto3Chunks()
-    {
-        int oldChunSize = TextProcessor.ChunkSize;
-        int oldOverlap = TextProcessor.Overlap;
-        TextProcessor.ChunkSize = 9;
-        TextProcessor.Overlap = 2;
-
-        await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None);
-
-        openAIClientMock.Verify(
-            client => client.GetEmbeddingsAsync(It.IsAny<EmbeddingsOptions>(), It.IsAny<CancellationToken>()),
-            Times.Exactly(3));
-        TextProcessor.ChunkSize = oldChunSize;
-        TextProcessor.Overlap = oldOverlap;
-    }
-
-    [Fact]
-    public async Task EmbedAsync_OverlapsChunks()
-    {
-        int oldChunSize = TextProcessor.ChunkSize;
-        int oldOverlap = TextProcessor.Overlap;
-        TextProcessor.ChunkSize = 11;
-        TextProcessor.Overlap = 3;
-
-        await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None);
-
-        openAIClientMock.Verify(
-            client => client.GetEmbeddingsAsync(It.Is<EmbeddingsOptions>(eo => eo.Input[0] == "Test docume"), It.IsAny<CancellationToken>()),
-            Times.Once);
-        openAIClientMock.Verify(
-            client => client.GetEmbeddingsAsync(It.Is<EmbeddingsOptions>(eo => eo.Input[0] == "ument text"), It.IsAny<CancellationToken>()),
-            Times.Once);
-        TextProcessor.ChunkSize = oldChunSize;
-        TextProcessor.Overlap = oldOverlap;
-    }
-
-    [Fact]
-    public async Task EmbedAsync_ReturnsEmbeddedTextChunks()
-    {
-        TextChunk expectedChunk = new()
-        {
-            Content = "Test document text",
-            Embedding = TestEmbeddings.CreateEmbeddingArray(),
-            DocumentReference = "file"
-        };
-        docTypeMock
-            .Setup(doc => doc.GetTextAsync(It.IsAny<Stream>(), It.IsAny<string>()))
-            .ReturnsAsync("Test document text");
-
-        List<TextChunk> textChunks = await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None);
-
-        TextChunk actualChunk = textChunks.First();
-        Assert.Equal(expectedChunk.Content, actualChunk.Content);
-        Assert.Equal(expectedChunk.DocumentReference, actualChunk.DocumentReference);
-        float[] expectedEmbedding = expectedChunk.Embedding.ToArray();
-        float[] actualEmbedding = actualChunk.Embedding.ToArray();
-        Assert.Equal(expectedEmbedding.Length, actualEmbedding.Length);
-        for (int i = 0; i < expectedEmbedding.Length; i++)
-            Assert.Equal(expectedEmbedding[i], actualEmbedding[i]);
-    }
-
-    [Fact]
-    public async Task EmbedAsync_WithRequestFailedException_LogsAnError()
-    {
-        openAIClientMock
-            .Setup(client => client.GetEmbeddingsAsync(It.IsAny<EmbeddingsOptions>(), It.IsAny<CancellationToken>()))
-            .Throws(new RequestFailedException("Unauthorized"));
-
-        await Assert.ThrowsAsync<RequestFailedException>(async () => 
-            await azureOpenAIRag.EmbedAsync("Test document text", docTypeMock.Object, "file", CancellationToken.None));
-
-        logMock
-            .Verify(
-                l => l.Log(
-                    LogLevel.Error,
-                    It.IsAny<EventId>(),
-                    It.IsAny<It.IsAnyType>(),
-                    It.IsAny<Exception>(),
-                    (Func<It.IsAnyType, Exception?, string>)It.IsAny<object>()),
-                Times.Once);
+            azSearchSvcMock.Object, docHashRepMock.Object, embedMock.Object, llmClientFactMock.Object, logMock.Object);
     }
 
     [Fact]
